+++ object/boolean_test.go
@@ -5,6 +5,20 @@ import (
 	"testing"
 )
 
+func testBooleanObject(t *testing.T, obj object.Object, expected bool) bool {
+	result, ok := obj.(*object.Boolean)
+	if !ok {
+		t.Errorf("object is not Boolean. got=%T (%+v)", obj, obj)
+		return false
+	}
+	if result.Value != expected {
+		t.Errorf("object has wrong value. got=%T, want=%t", result.Value, expected)
+		return false
+	}
+
+	return true
+}
+
 func TestBooleanObjectMethods(t *testing.T) {
 	tests := []inputTestCase{
 		{`true.plz_s()`, "true"},
