+++ object/builtin.go
@@ -1,0 +1,9 @@
+package object
+
+type Builtin struct {
+	Fn func(args ...Object) Object
+}
+
+func (b *Builtin) Type() ObjectType                                                   { return BUILTIN_OBJ }
+func (b *Builtin) Inspect() string                                                    { return "builtin function" }
+func (b *Builtin) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
