+++ object/error.go
@@ -1,0 +1,9 @@
+package object
+
+type Error struct {
+	Message string
+}
+
+func (e *Error) Type() ObjectType                                                   { return ERROR_OBJ }
+func (e *Error) Inspect() string                                                    { return "ERROR: " + e.Message }
+func (e *Error) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
