+++ object/integer_test.go
@@ -5,6 +5,20 @@ import (
 	"testing"
 )
 
+func testIntegerObject(t *testing.T, obj object.Object, expected int64) bool {
+	result, ok := obj.(*object.Integer)
+	if !ok {
+		t.Errorf("object is not Integer. got=%T (%+v)", obj, obj)
+		return false
+	}
+	if result.Value != expected {
+		t.Errorf("object has wrong value. got=%d, want=%d", result.Value, expected)
+		return false
+	}
+
+	return true
+}
+
 func TestIntegerObjectMethods(t *testing.T) {
 	tests := []inputTestCase{
 		{`2.plz_s()`, "2"},
