+++ object/object.go
@@ -1,11 +1,8 @@
 package object
 
 import (
-	"bytes"
 	"fmt"
 	"strings"
-
-	"github.com/flipez/rocket-lang/ast"
 )
 
 type ObjectType string
@@ -14,6 +11,15 @@ type Object interface {
 	Type() ObjectType
 	Inspect() string
 	InvokeMethod(method string, env Environment, args ...Object) Object
+}
+
+type Iterable interface {
+	Reset()
+	Next() (Object, Object, bool)
+}
+
+type Hashable interface {
+	HashKey() HashKey
 }
 
 const (
@@ -29,66 +35,6 @@ const (
 	HASH_OBJ         = "HASH"
 	FILE_OBJ         = "FILE"
 )
-
-type BuiltinFunction func(args ...Object) Object
-
-type Builtin struct {
-	Fn BuiltinFunction
-}
-
-func (b *Builtin) Type() ObjectType                                                   { return BUILTIN_OBJ }
-func (b *Builtin) Inspect() string                                                    { return "builtin function" }
-func (b *Builtin) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
-
-type Function struct {
-	Parameters []*ast.Identifier
-	Body       *ast.BlockStatement
-	Env        *Environment
-}
-
-func (f *Function) Type() ObjectType { return FUNCTION_OBJ }
-func (f *Function) Inspect() string {
-	var out bytes.Buffer
-
-	params := []string{}
-	for _, p := range f.Parameters {
-		params = append(params, p.String())
-	}
-
-	out.WriteString("fn")
-	out.WriteString("(")
-	out.WriteString(strings.Join(params, ", "))
-	out.WriteString(") {\n")
-	out.WriteString(f.Body.String())
-	out.WriteString("\n}")
-
-	return out.String()
-}
-func (f *Function) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
-
-type Error struct {
-	Message string
-}
-
-func (e *Error) Type() ObjectType                                                   { return ERROR_OBJ }
-func (e *Error) Inspect() string                                                    { return "ERROR: " + e.Message }
-func (e *Error) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
-
-type Null struct{}
-
-func (n *Null) Type() ObjectType                                                   { return NULL_OBJ }
-func (n *Null) Inspect() string                                                    { return "null" }
-func (n *Null) InvokeMethod(method string, env Environment, args ...Object) Object { return nil }
-
-type ReturnValue struct {
-	Value Object
-}
-
-func (rv *ReturnValue) Type() ObjectType { return RETURN_VALUE_OBJ }
-func (rv *ReturnValue) Inspect() string  { return rv.Value.Inspect() }
-func (rv *ReturnValue) InvokeMethod(method string, env Environment, args ...Object) Object {
-	return nil
-}
 
 type ObjectMethod struct {
 	argsOptional   bool
@@ -217,11 +163,6 @@ func objectMethodLookup(o Object, method string, args []Object) Object {
 	}
 
 	return nil
-}
-
-type Iterable interface {
-	Reset()
-	Next() (Object, Object, bool)
 }
 
 func CompareObjects(ao, bo Object) bool {
