+++ object/object_test.go
@@ -9,6 +9,11 @@ import (
 	"testing"
 )
 
+type inputTestCase struct {
+	input    string
+	expected interface{}
+}
+
 func testEval(input string) object.Object {
 	l := lexer.New(input)
 	p := parser.New(l)
@@ -16,53 +21,6 @@ func testEval(input string) object.Object {
 	env := object.NewEnvironment()
 
 	return evaluator.Eval(program, env)
-}
-
-func testIntegerObject(t *testing.T, obj object.Object, expected int64) bool {
-	result, ok := obj.(*object.Integer)
-	if !ok {
-		t.Errorf("object is not Integer. got=%T (%+v)", obj, obj)
-		return false
-	}
-	if result.Value != expected {
-		t.Errorf("object has wrong value. got=%d, want=%d", result.Value, expected)
-		return false
-	}
-
-	return true
-}
-
-func testStringObject(t *testing.T, obj object.Object, expected string) bool {
-	result, ok := obj.(*object.String)
-	if !ok {
-		t.Errorf("obj is not String. got=%T(%+v)", obj, obj)
-		return false
-	}
-	if result.Value != expected {
-		t.Errorf("object has wrong value. got=%s, want=%s",
-			result.Value, expected)
-		return false
-	}
-	return true
-}
-
-func testBooleanObject(t *testing.T, obj object.Object, expected bool) bool {
-	result, ok := obj.(*object.Boolean)
-	if !ok {
-		t.Errorf("object is not Boolean. got=%T (%+v)", obj, obj)
-		return false
-	}
-	if result.Value != expected {
-		t.Errorf("object has wrong value. got=%T, want=%t", result.Value, expected)
-		return false
-	}
-
-	return true
-}
-
-type inputTestCase struct {
-	input    string
-	expected interface{}
 }
 
 func testInput(t *testing.T, tests []inputTestCase) {
@@ -94,76 +52,6 @@ func testInput(t *testing.T, tests []inputTestCase) {
 			}
 		case bool:
 			testBooleanObject(t, evaluated, expected)
-		}
-	}
-}
-
-func TestReturnValue(t *testing.T) {
-	rv := &object.ReturnValue{Value: &object.String{Value: "a"}}
-
-	if rv.Type() != object.RETURN_VALUE_OBJ {
-		t.Errorf("returnValue.Type() returns wrong type")
-	}
-	if rv.Inspect() != "a" {
-		t.Errorf("returnValue.Inspect() returns wrong type")
-	}
-}
-
-func TestNullType(t *testing.T) {
-	n := &object.Null{}
-
-	if n.Type() != object.NULL_OBJ {
-		t.Errorf("null.Type() returns wrong type")
-	}
-	if n.Inspect() != "null" {
-		t.Errorf("null.Inspect() returns wrong type")
-	}
-}
-func TestNullObjectMethods(t *testing.T) {
-	tests := []inputTestCase{
-		{`[1][1].nope()`, "Failed to invoke method: nope"},
-	}
-
-	testInput(t, tests)
-}
-
-func TestErrorType(t *testing.T) {
-	err := &object.Error{Message: "test"}
-
-	if err.Type() != object.ERROR_OBJ {
-		t.Errorf("error.Type() returns wrong type")
-	}
-	if err.Inspect() != "ERROR: test" {
-		t.Errorf("error.Inspect() returns wrong type")
-	}
-}
-func TestErrorObjectMethods(t *testing.T) {
-	tests := []inputTestCase{
-		{`[1][1].nope().nope()`, "Failed to invoke method: nope"},
-	}
-
-	testInput(t, tests)
-}
-
-func TestFunctionObjectMethods(t *testing.T) {
-	tests := []inputTestCase{
-		{`fn(){}.nope()`, "Failed to invoke method: nope"},
-	}
-
-	testInput(t, tests)
-}
-func TestFunctionType(t *testing.T) {
-	tests := []inputTestCase{
-		{"fn(){}", "fn() {\n\n}"},
-		{"fn(a){puts(a)}", "fn(a) {\nputs(a)\n}"},
-	}
-
-	for _, tt := range tests {
-		fn := testEval(tt.input).(*object.Function)
-		fnInspect := fn.Inspect()
-
-		if fnInspect != tt.expected {
-			t.Errorf("wrong string. expected=%#v, got=%#v", tt.expected, fnInspect)
 		}
 	}
 }
