+++ object/return_value.go
@@ -1,0 +1,11 @@
+package object
+
+type ReturnValue struct {
+	Value Object
+}
+
+func (rv *ReturnValue) Type() ObjectType { return RETURN_VALUE_OBJ }
+func (rv *ReturnValue) Inspect() string  { return rv.Value.Inspect() }
+func (rv *ReturnValue) InvokeMethod(method string, env Environment, args ...Object) Object {
+	return nil
+}
