+++ object/return_value_test.go
@@ -1,0 +1,18 @@
+package object_test
+
+import (
+	"testing"
+
+	"github.com/flipez/rocket-lang/object"
+)
+
+func TestReturnValue(t *testing.T) {
+	rv := &object.ReturnValue{Value: &object.String{Value: "a"}}
+
+	if rv.Type() != object.RETURN_VALUE_OBJ {
+		t.Errorf("returnValue.Type() returns wrong type")
+	}
+	if rv.Inspect() != "a" {
+		t.Errorf("returnValue.Inspect() returns wrong type")
+	}
+}
